requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
PyYAML>=6.0
Django>=4.2
